Centralize common getCertificate() behavior
