+++ api/lib/domain/usecases/index.js
@@ -159,11 +159,11 @@ module.exports = injectDependencies({
   getSession: require('./get-session'),
   getSessionCertificationCandidates: require('./get-session-certification-candidates'),
   getSessionCertificationReports: require('./get-session-certification-reports'),
-  getCertificationByVerificationCode: require('./get-shared-certification'),
+  getCertificationByVerificationCode: require('./certificate/get-shared-certification'),
   getUserCampaignParticipationToCampaign: require('./get-user-campaign-participation-to-campaign'),
   getUserProfileSharedForCampaign: require('./get-user-profile-shared-for-campaign'),
   getUserCertificationCenterMemberships: require('./get-user-certification-center-memberships'),
-  getUserCertificationWithResultTree: require('./get-user-certification-with-result-tree'),
+  getUserCertificationWithResultTree: require('./certificate/get-user-certification-with-result-tree'),
   isUserCertifiable: require('./is-user-certifiable'),
   getUserDetailsForAdmin: require('./get-user-details-for-admin'),
   getUserPixScore: require('./get-user-pix-score'),
